requirements.txt
streamlit
plotly
pandas
openai
python-dotenv
tiktoken
